+++ src/components/icons/about.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function IconAbout () {
+  return (
+    <div>A</div>
+  )
+}
+
+export default IconAbout