+++ src/components/icons/settings.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function IconSettings () {
+  return (
+    <div>S</div>
+  )
+}
+
+export default IconSettings