+++ src/pages/not-found/index.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import './_index.scss'
+
+function NotFoundPage () {
+  return (
+    <div className="page -not-found">
+      <div>404</div>
+      <div>
+        <a href="/">home page</a>
+      </div>
+    </div>
+  )
+}
+
+export default NotFoundPage