+++ src/pages/settings/general.tsx
@@ -3,12 +3,19 @@ import React from 'react'
 function SettingsAppearancePage () {
   return (
     <div className="page -settings-appearance">
-      <div>
+      <section>
         <div>Color theme:</div>
-        <select disabled={true}>
+        <select defaultValue="dark" disabled={true}>
           <option value="dark">Dark</option>
         </select>
-      </div>
+      </section>
+
+      <section>
+        <div>Language:</div>
+        <select defaultValue="en" disabled={true}>
+          <option value="en">English</option>
+        </select>
+      </section>
     </div>
   )
 }