+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+### 1.2.1 (2021-06-23)
+* Set security code modal updated.
+* Icon wrapper component added.
+* Some content loaders added in the app.
+
+
 ### 1.2.0 (2021-06-22)
 * Automatic website logo checker added.
 * Primary website image color as image background.